Fix namespace of iFDO report generators
The namespace should start either with "Image" or "Video".
