+++ src/Support/Reports/Projects/VideoIfdoReportGenerator.php
@@ -1,12 +1,11 @@
 <?php
 
-namespace Biigle\Modules\Reports\Support\Reports\Projects\Ifdo;
+namespace Biigle\Modules\Reports\Support\Reports\Projects;
 
-use Biigle\Modules\Reports\Support\Reports\Projects\ProjectImageReportGenerator;
-use Biigle\Modules\Reports\Support\Reports\Volumes\ImageIfdoReportGenerator as ReportGenerator;
+use Biigle\Modules\Reports\Support\Reports\Volumes\VideoIfdoReportGenerator as ReportGenerator;
 use Exception;
 
-class ImageIfdoReportGenerator extends ProjectImageReportGenerator
+class VideoIfdoReportGenerator extends ProjectVideoReportGenerator
 {
     /**
      * The class of the volume report to use for this project report.
@@ -20,14 +19,14 @@ class ImageIfdoReportGenerator extends ProjectImageReportGenerator
      *
      * @var string
      */
-    protected $name = 'image iFDO report';
+    protected $name = 'video iFDO report';
 
     /**
      * Name of the report for use as (part of) a filename.
      *
      * @var string
      */
-    protected $filename = 'image_ifdo_report';
+    protected $filename = 'video_ifdo_report';
 
     /**
      * Get sources for the sub-reports that should be generated for this project.
