+++ tests/Support/Reports/Volumes/VideoIfdoReportGeneratorTest.php
@@ -1,15 +1,15 @@
 <?php
 
-namespace Biigle\Tests\Modules\Reports\Support\Reports\Volumes\Ifdo;
+namespace Biigle\Tests\Modules\Reports\Support\Reports\Volumes;
 
 use Biigle\LabelSource;
 use Biigle\Shape;
-use Biigle\Modules\Reports\Support\Reports\Volumes\Ifdo\ImageIfdoReportGenerator;
+use Biigle\Modules\Reports\Support\Reports\Volumes\VideoIfdoReportGenerator;
 use Biigle\Modules\Reports\Volume;
-use Biigle\Tests\ImageAnnotationLabelTest;
-use Biigle\Tests\ImageAnnotationTest;
-use Biigle\Tests\ImageLabelTest;
-use Biigle\Tests\ImageTest;
+use Biigle\Tests\VideoAnnotationLabelTest;
+use Biigle\Tests\VideoAnnotationTest;
+use Biigle\Tests\VideoLabelTest;
+use Biigle\Tests\VideoTest;
 use Biigle\Tests\LabelTest;
 use Biigle\Tests\UserTest;
 use Biigle\Tests\VolumeTest;
@@ -17,13 +17,13 @@ use Exception;
 use Storage;
 use TestCase;
 
-class ImageIfdoReportGeneratorTest extends TestCase
+class VideoIfdoReportGeneratorTest extends TestCase
 {
     public function testProperties()
     {
-        $generator = new ImageIfdoReportGenerator;
-        $this->assertEquals('image iFDO report', $generator->getName());
-        $this->assertEquals('image_ifdo_report', $generator->getFilename());
+        $generator = new VideoIfdoReportGenerator;
+        $this->assertEquals('video iFDO report', $generator->getName());
+        $this->assertEquals('video_ifdo_report', $generator->getFilename());
         $this->assertStringEndsWith('.yaml', $generator->getFullFilename());
     }
 
@@ -34,7 +34,7 @@ class ImageIfdoReportGeneratorTest extends TestCase
                 'image-set-handle' => '20.500.12085/test-example',
                 'image-set-name' => 'My Cool Volume',
                 'image-set-uuid' => 'd7546c4b-307f-4d42-8554-33236c577450',
-                'image-acquisition' => 'image',
+                'image-acquisition' => 'video',
             ],
         ], $merge);
 
@@ -53,19 +53,20 @@ class ImageIfdoReportGeneratorTest extends TestCase
         $label = LabelTest::create();
         $user = UserTest::create();
 
-        $image = ImageTest::create(['volume_id' => $volume->id]);
-        $a = ImageAnnotationTest::create([
-            'image_id' => $image->id,
-            'points' => [150, 150],
+        $video = VideoTest::create(['volume_id' => $volume->id]);
+        $a = VideoAnnotationTest::create([
+            'video_id' => $video->id,
+            'points' => [[150, 150], [200, 200]],
+            'frames' => [100.0, 200.0],
             'shape_id' => Shape::pointId(),
         ]);
-        $al = ImageAnnotationLabelTest::create([
+        $al = VideoAnnotationLabelTest::create([
             'label_id' => $label->id,
             'user_id' => $user->id,
             'annotation_id' => $a->id,
         ]);
 
-        $generator = new ImageIfdoReportGeneratorStub;
+        $generator = new VideoIfdoReportGeneratorStub;
         $generator->setSource($volume);
         $generator->generateReport('my/path');
 
@@ -74,7 +75,7 @@ class ImageIfdoReportGeneratorTest extends TestCase
                 'image-set-handle' => '20.500.12085/test-example',
                 'image-set-name' => 'My Cool Volume',
                 'image-set-uuid' => 'd7546c4b-307f-4d42-8554-33236c577450',
-                'image-acquisition' => 'image',
+                'image-acquisition' => 'video',
                 'image-annotation-creators' => [
                     [
                         'id' => $user->uuid,
@@ -90,11 +91,12 @@ class ImageIfdoReportGeneratorTest extends TestCase
                 ],
             ],
             'image-set-items' => [
-                $al->annotation->image->filename => [
+                $al->annotation->video->filename => [[
                     'image-annotation-geometry-types' => ['single-pixel'],
                     'image-annotations' => [
                         [
-                            'coordinates' => $al->annotation->points,
+                            'coordinates' => [[150, 150], [200, 200]],
+                            'frames' => [100.0, 200.0],
                             'labels' => [
                                 [
                                     'label' => $al->label_id,
@@ -105,7 +107,7 @@ class ImageIfdoReportGeneratorTest extends TestCase
                             ],
                         ],
                     ],
-                ],
+                ]],
             ],
         ];
 
@@ -120,25 +122,26 @@ class ImageIfdoReportGeneratorTest extends TestCase
         $label2 = LabelTest::create();
         $user = UserTest::create();
 
-        $image = ImageTest::create(['volume_id' => $volume->id]);
-        $a = ImageAnnotationTest::create([
-            'image_id' => $image->id,
-            'points' => [150, 150],
+        $video = VideoTest::create(['volume_id' => $volume->id]);
+        $a = VideoAnnotationTest::create([
+            'video_id' => $video->id,
+            'points' => [[150, 150]],
+            'frames' => [100.0],
             'shape_id' => Shape::pointId(),
         ]);
-        $al = ImageAnnotationLabelTest::create([
+        $al = VideoAnnotationLabelTest::create([
             'label_id' => $label->id,
             'user_id' => $user->id,
             'annotation_id' => $a->id,
         ]);
 
-        $al2 = ImageAnnotationLabelTest::create([
+        $al2 = VideoAnnotationLabelTest::create([
             'label_id' => $label2->id,
             'user_id' => $user->id,
             'annotation_id' => $a->id,
         ]);
 
-        $generator = new ImageIfdoReportGeneratorStub;
+        $generator = new VideoIfdoReportGeneratorStub;
         $generator->setSource($volume);
         $generator->generateReport('my/path');
 
@@ -147,7 +150,7 @@ class ImageIfdoReportGeneratorTest extends TestCase
                 'image-set-handle' => '20.500.12085/test-example',
                 'image-set-name' => 'My Cool Volume',
                 'image-set-uuid' => 'd7546c4b-307f-4d42-8554-33236c577450',
-                'image-acquisition' => 'image',
+                'image-acquisition' => 'video',
                 'image-annotation-creators' => [
                     [
                         'id' => $user->uuid,
@@ -167,11 +170,12 @@ class ImageIfdoReportGeneratorTest extends TestCase
                 ],
             ],
             'image-set-items' => [
-                $al->annotation->image->filename => [
+                $al->annotation->video->filename => [[
                     'image-annotation-geometry-types' => ['single-pixel'],
                     'image-annotations' => [
                         [
-                            'coordinates' => $al->annotation->points,
+                            'coordinates' => [[150, 150]],
+                            'frames' => [100.0],
                             'labels' => [
                                 [
                                     'label' => $al->label_id,
@@ -188,7 +192,7 @@ class ImageIfdoReportGeneratorTest extends TestCase
                             ],
                         ],
                     ],
-                ],
+                ]],
             ],
         ];
 
@@ -199,9 +203,9 @@ class ImageIfdoReportGeneratorTest extends TestCase
     {
         [$volume, $ifdo] = $this->setUpIfdo();
 
-        $image = ImageTest::create(['volume_id' => $volume->id]);
+        $video = VideoTest::create(['volume_id' => $volume->id]);
 
-        $generator = new ImageIfdoReportGeneratorStub;
+        $generator = new VideoIfdoReportGeneratorStub;
         $generator->setSource($volume);
         $generator->generateReport('my/path');
 
@@ -210,32 +214,32 @@ class ImageIfdoReportGeneratorTest extends TestCase
                 'image-set-handle' => '20.500.12085/test-example',
                 'image-set-name' => 'My Cool Volume',
                 'image-set-uuid' => 'd7546c4b-307f-4d42-8554-33236c577450',
-                'image-acquisition' => 'image',
+                'image-acquisition' => 'video',
             ],
             'image-set-items' => [
-                $image->filename => [],
+                $video->filename => [],
             ],
         ];
 
         $this->assertEquals($expect, $generator->yaml);
     }
 
-    public function testGenerateReportImageLabels()
+    public function testGenerateReportVideoLabels()
     {
         [$volume, $ifdo] = $this->setUpIfdo();
 
         $label = LabelTest::create();
         $user = UserTest::create();
 
-        $il = ImageLabelTest::create([
+        $vl = VideoLabelTest::create([
             'label_id' => $label->id,
             'user_id' => $user->id,
         ]);
 
-        $il->image->volume_id = $volume->id;
-        $il->image->save();
+        $vl->video->volume_id = $volume->id;
+        $vl->video->save();
 
-        $generator = new ImageIfdoReportGeneratorStub;
+        $generator = new VideoIfdoReportGeneratorStub;
         $generator->setSource($volume);
         $generator->generateReport('my/path');
 
@@ -244,7 +248,7 @@ class ImageIfdoReportGeneratorTest extends TestCase
                 'image-set-handle' => '20.500.12085/test-example',
                 'image-set-name' => 'My Cool Volume',
                 'image-set-uuid' => 'd7546c4b-307f-4d42-8554-33236c577450',
-                'image-acquisition' => 'image',
+                'image-acquisition' => 'video',
                 'image-annotation-creators' => [
                     [
                         'id' => $user->uuid,
@@ -254,27 +258,27 @@ class ImageIfdoReportGeneratorTest extends TestCase
                 ],
                 'image-annotation-labels' => [
                     [
-                        'id' => $il->label_id,
-                        'name' => $il->label->name,
+                        'id' => $vl->label_id,
+                        'name' => $vl->label->name,
                     ],
                 ],
             ],
             'image-set-items' => [
-                $il->image->filename => [
+                $vl->video->filename => [[
                     'image-annotation-geometry-types' => ['whole-image'],
                     'image-annotations' => [
                         [
                             'coordinates' => [],
                             'labels' => [
                                 [
-                                    'label' => $il->label_id,
+                                    'label' => $vl->label_id,
                                     'annotator' => $user->uuid,
-                                    'created-at' => (string) $il->created_at,
+                                    'created-at' => (string) $vl->created_at,
                                 ],
                             ],
                         ],
                     ],
-                ],
+                ]],
             ],
         ];
 
@@ -286,7 +290,7 @@ class ImageIfdoReportGeneratorTest extends TestCase
         $label = LabelTest::create();
         $user = UserTest::create();
 
-        $image = ImageTest::create();
+        $video = VideoTest::create();
 
         $merge = [
             'image-set-header' => [
@@ -305,7 +309,7 @@ class ImageIfdoReportGeneratorTest extends TestCase
                 ],
             ],
             'image-set-items' => [
-                $image->filename => [
+                $video->filename => [[
                     'image-area-square-meter' => 5.5,
                     'image-annotation-geometry-types' => ['bounding-box'],
                     'image-annotations' => [
@@ -321,27 +325,28 @@ class ImageIfdoReportGeneratorTest extends TestCase
                             ],
                         ],
                     ],
-                ],
+                ]],
             ],
         ];
 
         [$volume, $ifdo] = $this->setUpIfdo($merge);
 
-        $a = ImageAnnotationTest::create([
-            'image_id' => $image->id,
-            'points' => [150, 150],
+        $a = VideoAnnotationTest::create([
+            'video_id' => $video->id,
+            'points' => [[150, 150]],
+            'frames' => [100.0],
             'shape_id' => Shape::pointId(),
         ]);
-        $al = ImageAnnotationLabelTest::create([
+        $al = VideoAnnotationLabelTest::create([
             'label_id' => $label->id,
             'user_id' => $user->id,
             'annotation_id' => $a->id,
         ]);
 
-        $image->volume_id = $volume->id;
-        $image->save();
+        $video->volume_id = $volume->id;
+        $video->save();
 
-        $generator = new ImageIfdoReportGeneratorStub;
+        $generator = new VideoIfdoReportGeneratorStub;
         $generator->setSource($volume);
         $generator->generateReport('my/path');
 
@@ -350,7 +355,7 @@ class ImageIfdoReportGeneratorTest extends TestCase
                 'image-set-handle' => '20.500.12085/test-example',
                 'image-set-name' => 'My Cool Volume',
                 'image-set-uuid' => 'd7546c4b-307f-4d42-8554-33236c577450',
-                'image-acquisition' => 'image',
+                'image-acquisition' => 'video',
                 'image-annotation-creators' => [
                     [
                         'id' => '123abc',
@@ -375,7 +380,7 @@ class ImageIfdoReportGeneratorTest extends TestCase
                 ],
             ],
             'image-set-items' => [
-                $al->annotation->image->filename => [
+                $al->annotation->video->filename => [[
                     'image-area-square-meter' => 5.5,
                     'image-annotation-geometry-types' => ['bounding-box', 'single-pixel'],
                     'image-annotations' => [
@@ -391,7 +396,8 @@ class ImageIfdoReportGeneratorTest extends TestCase
                             ],
                         ],
                         [
-                            'coordinates' => $al->annotation->points,
+                            'coordinates' => [[150, 150]],
+                            'frames' => [100.0],
                             'labels' => [
                                 [
                                     'label' => $al->label_id,
@@ -402,46 +408,51 @@ class ImageIfdoReportGeneratorTest extends TestCase
                             ],
                         ],
                     ],
-                ],
+                ]],
             ],
         ];
 
         $this->assertEquals($expect, $generator->yaml);
     }
 
-    public function testGenerateReportMergeImageSetItemsArray()
+    public function testGenerateReportMergeImageSetItemsMultiple()
     {
         $label = LabelTest::create();
         $user = UserTest::create();
 
-        $image = ImageTest::create();
+        $video = VideoTest::create();
 
         $merge = [
             'image-set-items' => [
-                // The double array here is the difference to testGenerateReportMergeImageSetItems!
-                $image->filename => [[
-                    'image-area-square-meter' => 5.5,
-                ]],
+                $video->filename => [
+                    [
+                        'image-area-square-meter' => 5.5,
+                    ],
+                    [
+                        'image-area-square-meter' => 6.0,
+                    ],
+                ],
             ],
         ];
 
         [$volume, $ifdo] = $this->setUpIfdo($merge);
 
-        $a = ImageAnnotationTest::create([
-            'image_id' => $image->id,
-            'points' => [150, 150],
+        $a = VideoAnnotationTest::create([
+            'video_id' => $video->id,
+            'points' => [[150, 150]],
+            'frames' => [100.0],
             'shape_id' => Shape::pointId(),
         ]);
-        $al = ImageAnnotationLabelTest::create([
+        $al = VideoAnnotationLabelTest::create([
             'label_id' => $label->id,
             'user_id' => $user->id,
             'annotation_id' => $a->id,
         ]);
 
-        $image->volume_id = $volume->id;
-        $image->save();
+        $video->volume_id = $volume->id;
+        $video->save();
 
-        $generator = new ImageIfdoReportGeneratorStub;
+        $generator = new VideoIfdoReportGeneratorStub;
         $generator->setSource($volume);
         $generator->generateReport('my/path');
 
@@ -450,7 +461,7 @@ class ImageIfdoReportGeneratorTest extends TestCase
                 'image-set-handle' => '20.500.12085/test-example',
                 'image-set-name' => 'My Cool Volume',
                 'image-set-uuid' => 'd7546c4b-307f-4d42-8554-33236c577450',
-                'image-acquisition' => 'image',
+                'image-acquisition' => 'video',
                 'image-annotation-creators' => [
                     [
                         'id' => $user->uuid,
@@ -466,104 +477,27 @@ class ImageIfdoReportGeneratorTest extends TestCase
                 ],
             ],
             'image-set-items' => [
-                $al->annotation->image->filename => [[
-                    'image-area-square-meter' => 5.5,
-                    'image-annotation-geometry-types' => ['single-pixel'],
-                    'image-annotations' => [
-                        [
-                            'coordinates' => $al->annotation->points,
-                            'labels' => [
-                                [
-                                    'label' => $al->label_id,
-                                    'annotator' => $user->uuid,
-                                    'confidence' => $al->confidence,
-                                    'created-at' => (string) $al->created_at,
+                $al->annotation->video->filename => [
+                    [
+                        'image-area-square-meter' => 5.5,
+                        'image-annotation-geometry-types' => ['single-pixel'],
+                        'image-annotations' => [
+                            [
+                                'coordinates' => [[150, 150]],
+                                'frames' => [100.0],
+                                'labels' => [
+                                    [
+                                        'label' => $al->label_id,
+                                        'annotator' => $user->uuid,
+                                        'confidence' => $al->confidence,
+                                        'created-at' => (string) $al->created_at,
+                                    ],
                                 ],
                             ],
                         ],
                     ],
-                ]],
-            ],
-        ];
-
-        $this->assertEquals($expect, $generator->yaml);
-    }
-
-    public function testGenerateReportRestrictToExportArea()
-    {
-        [$volume, $ifdo] = $this->setUpIfdo();
-
-        $volume = Volume::convert($volume);
-        $volume->exportArea = [100, 100, 200, 200];
-        $volume->save();
-
-        $label = LabelTest::create();
-        $user = UserTest::create();
-
-        $image = ImageTest::create(['volume_id' => $volume->id]);
-        $a1 = ImageAnnotationTest::create([
-            'image_id' => $image->id,
-            'shape_id' => Shape::pointId(),
-            'points' => [150, 150],
-        ]);
-        $al1 = ImageAnnotationLabelTest::create([
-            'label_id' => $label->id,
-            'user_id' => $user->id,
-            'annotation_id' => $a1->id,
-        ]);
-
-        $a2 = ImageAnnotationTest::create([
-            'image_id' => $image->id,
-            'shape_id' => Shape::pointId(),
-            'points' => [50, 50],
-        ]);
-        $al2 = ImageAnnotationLabelTest::create([
-            'label_id' => $label->id,
-            'user_id' => $user->id,
-            'annotation_id' => $a2->id,
-        ]);
-
-        $generator = new ImageIfdoReportGeneratorStub([
-            'exportArea' => true,
-        ]);
-        $generator->setSource($volume);
-        $generator->generateReport('my/path');
-
-        $expect = [
-            'image-set-header' => [
-                'image-set-handle' => '20.500.12085/test-example',
-                'image-set-name' => 'My Cool Volume',
-                'image-set-uuid' => 'd7546c4b-307f-4d42-8554-33236c577450',
-                'image-acquisition' => 'image',
-                'image-annotation-creators' => [
                     [
-                        'id' => $user->uuid,
-                        'name' => "{$user->firstname} {$user->lastname}",
-                        'type' => 'expert',
-                    ],
-                ],
-                'image-annotation-labels' => [
-                    [
-                        'id' => $al1->label_id,
-                        'name' => $al1->label->name,
-                    ],
-                ],
-            ],
-            'image-set-items' => [
-                $image->filename => [
-                    'image-annotation-geometry-types' => ['single-pixel'],
-                    'image-annotations' => [
-                        [
-                            'coordinates' => $a1->points,
-                            'labels' => [
-                                [
-                                    'label' => $al1->label_id,
-                                    'annotator' => $user->uuid,
-                                    'confidence' => $al1->confidence,
-                                    'created-at' => (string) $al1->created_at,
-                                ],
-                            ],
-                        ],
+                        'image-area-square-meter' => 6.0,
                     ],
                 ],
             ],
@@ -580,25 +514,26 @@ class ImageIfdoReportGeneratorTest extends TestCase
         $label2 = LabelTest::create();
         $user = UserTest::create();
 
-        $image = ImageTest::create(['volume_id' => $volume->id]);
-        $a = ImageAnnotationTest::create([
-            'image_id' => $image->id,
-            'points' => [150, 150],
+        $video = VideoTest::create(['volume_id' => $volume->id]);
+        $a = VideoAnnotationTest::create([
+            'video_id' => $video->id,
+            'points' => [[150, 150]],
+            'frames' => [100.0],
             'shape_id' => Shape::pointId(),
         ]);
-        $al = ImageAnnotationLabelTest::create([
+        $al = VideoAnnotationLabelTest::create([
             'label_id' => $label->id,
             'user_id' => $user->id,
             'annotation_id' => $a->id,
         ]);
 
-        $al2 = ImageAnnotationLabelTest::create([
+        $al2 = VideoAnnotationLabelTest::create([
             'label_id' => $label2->id,
             'user_id' => $user->id,
             'annotation_id' => $a->id,
         ]);
 
-        $generator = new ImageIfdoReportGeneratorStub([
+        $generator = new VideoIfdoReportGeneratorStub([
             'newestLabel' => true,
         ]);
         $generator->setSource($volume);
@@ -609,7 +544,7 @@ class ImageIfdoReportGeneratorTest extends TestCase
                 'image-set-handle' => '20.500.12085/test-example',
                 'image-set-name' => 'My Cool Volume',
                 'image-set-uuid' => 'd7546c4b-307f-4d42-8554-33236c577450',
-                'image-acquisition' => 'image',
+                'image-acquisition' => 'video',
                 'image-annotation-creators' => [
                     [
                         'id' => $user->uuid,
@@ -625,11 +560,12 @@ class ImageIfdoReportGeneratorTest extends TestCase
                 ],
             ],
             'image-set-items' => [
-                $al->annotation->image->filename => [
+                $al->annotation->video->filename => [[
                     'image-annotation-geometry-types' => ['single-pixel'],
                     'image-annotations' => [
                         [
-                            'coordinates' => $al->annotation->points,
+                            'coordinates' => [[150, 150]],
+                            'frames' => [100.0],
                             'labels' => [
                                 [
                                     'label' => $al2->label_id,
@@ -640,7 +576,7 @@ class ImageIfdoReportGeneratorTest extends TestCase
                             ],
                         ],
                     ],
-                ],
+                ]],
             ],
         ];
 
@@ -655,50 +591,50 @@ class ImageIfdoReportGeneratorTest extends TestCase
         $label2 = LabelTest::create();
         $user = UserTest::create();
 
-        $image1 = ImageTest::create([
+        $video1 = VideoTest::create([
             'volume_id' => $volume->id,
             'filename' => 'img1.jpg',
         ]);
-        $a1 = ImageAnnotationTest::create([
-            'image_id' => $image1->id,
+        $a1 = VideoAnnotationTest::create([
+            'video_id' => $video1->id,
             'shape_id' => Shape::pointId(),
         ]);
-        $al = ImageAnnotationLabelTest::create([
+        $al = VideoAnnotationLabelTest::create([
             'label_id' => $label->id,
             'user_id' => $user->id,
             'annotation_id' => $a1->id,
         ]);
 
-        $il1 = ImageLabelTest::create([
-            'image_id' => $image1->id,
+        $vl1 = VideoLabelTest::create([
+            'video_id' => $video1->id,
             'label_id' => $label->id,
             'user_id' => $user->id,
         ]);
 
-        $image2 = ImageTest::create([
+        $video2 = VideoTest::create([
             'volume_id' => $volume->id,
             'filename' => 'img2.jpg',
         ]);
-        $a2 = ImageAnnotationTest::create([
-            'image_id' => $image2->id,
+        $a2 = VideoAnnotationTest::create([
+            'video_id' => $video2->id,
             'shape_id' => Shape::pointId(),
         ]);
-        $al2 = ImageAnnotationLabelTest::create([
+        $al2 = VideoAnnotationLabelTest::create([
             'label_id' => $label2->id,
             'user_id' => $user->id,
             'annotation_id' => $a2->id,
         ]);
 
-        $il2 = ImageLabelTest::create([
-            'image_id' => $image2->id,
+        $vl2 = VideoLabelTest::create([
+            'video_id' => $video2->id,
             'label_id' => $label2->id,
             'user_id' => $user->id,
         ]);
 
-        $al2->annotation->image->volume_id = $volume->id;
-        $al2->annotation->image->save();
+        $al2->annotation->video->volume_id = $volume->id;
+        $al2->annotation->video->save();
 
-        $generator = new ImageIfdoReportGeneratorStub([
+        $generator = new VideoIfdoReportGeneratorStub([
             'onlyLabels' => [$label->id],
         ]);
         $generator->setSource($volume);
@@ -709,7 +645,7 @@ class ImageIfdoReportGeneratorTest extends TestCase
                 'image-set-handle' => '20.500.12085/test-example',
                 'image-set-name' => 'My Cool Volume',
                 'image-set-uuid' => 'd7546c4b-307f-4d42-8554-33236c577450',
-                'image-acquisition' => 'image',
+                'image-acquisition' => 'video',
                 'image-annotation-creators' => [
                     [
                         'id' => $user->uuid,
@@ -725,7 +661,7 @@ class ImageIfdoReportGeneratorTest extends TestCase
                 ],
             ],
             'image-set-items' => [
-                $image1->filename => [
+                $video1->filename => [[
                     'image-annotation-geometry-types' => [
                         'single-pixel',
                         'whole-image',
@@ -733,6 +669,7 @@ class ImageIfdoReportGeneratorTest extends TestCase
                     'image-annotations' => [
                         [
                             'coordinates' => $a1->points,
+                            'frames' => $a1->frames,
                             'labels' => [
                                 [
                                     'label' => $label->id,
@@ -748,13 +685,13 @@ class ImageIfdoReportGeneratorTest extends TestCase
                                 [
                                     'label' => $label->id,
                                     'annotator' => $user->uuid,
-                                    'created-at' => (string) $il1->created_at,
+                                    'created-at' => (string) $vl1->created_at,
                                 ],
                             ],
                         ],
                     ],
-                ],
-                $image2->filename => [],
+                ]],
+                $video2->filename => [],
             ],
         ];
 
@@ -764,7 +701,7 @@ class ImageIfdoReportGeneratorTest extends TestCase
     public function testGenerateReportNoIfdo()
     {
         $volume = VolumeTest::create();
-        $generator = new ImageIfdoReportGeneratorStub();
+        $generator = new VideoIfdoReportGeneratorStub();
         $generator->setSource($volume);
         $this->expectException(Exception::class);
         $generator->generateReport('my/path');
@@ -782,19 +719,20 @@ class ImageIfdoReportGeneratorTest extends TestCase
         ]);
         $user = UserTest::create();
 
-        $image = ImageTest::create(['volume_id' => $volume->id]);
-        $a = ImageAnnotationTest::create([
-            'image_id' => $image->id,
-            'points' => [150, 150],
+        $video = VideoTest::create(['volume_id' => $volume->id]);
+        $a = VideoAnnotationTest::create([
+            'video_id' => $video->id,
+            'points' => [[150, 150]],
+            'frames' => [100.0],
             'shape_id' => Shape::pointId(),
         ]);
-        $al = ImageAnnotationLabelTest::create([
+        $al = VideoAnnotationLabelTest::create([
             'label_id' => $label->id,
             'user_id' => $user->id,
             'annotation_id' => $a->id,
         ]);
 
-        $generator = new ImageIfdoReportGeneratorStub;
+        $generator = new VideoIfdoReportGeneratorStub;
         $generator->setSource($volume);
         $generator->generateReport('my/path');
 
@@ -803,7 +741,7 @@ class ImageIfdoReportGeneratorTest extends TestCase
                 'image-set-handle' => '20.500.12085/test-example',
                 'image-set-name' => 'My Cool Volume',
                 'image-set-uuid' => 'd7546c4b-307f-4d42-8554-33236c577450',
-                'image-acquisition' => 'image',
+                'image-acquisition' => 'video',
                 'image-annotation-creators' => [
                     [
                         'id' => $user->uuid,
@@ -819,11 +757,12 @@ class ImageIfdoReportGeneratorTest extends TestCase
                 ],
             ],
             'image-set-items' => [
-                $al->annotation->image->filename => [
+                $al->annotation->video->filename => [[
                     'image-annotation-geometry-types' => ['single-pixel'],
                     'image-annotations' => [
                         [
-                            'coordinates' => $al->annotation->points,
+                            'coordinates' => [[150, 150]],
+                            'frames' => [100.0],
                             'labels' => [
                                 [
                                     'label' => 'urn:lsid:marinespecies.org:taxname:123999',
@@ -834,7 +773,7 @@ class ImageIfdoReportGeneratorTest extends TestCase
                             ],
                         ],
                     ],
-                ],
+                ]],
             ],
         ];
 
@@ -846,7 +785,7 @@ class ImageIfdoReportGeneratorTest extends TestCase
         $label = LabelTest::create();
         $user = UserTest::create();
 
-        $image = ImageTest::create();
+        $video = VideoTest::create();
 
         $merge = [
             'image-set-header' => [
@@ -865,115 +804,7 @@ class ImageIfdoReportGeneratorTest extends TestCase
                 ],
             ],
             'image-set-items' => [
-                $image->filename => [
-                    'image-annotation-geometry-types' => ['bounding-box'],
-                    'image-annotations' => [
-                        [
-                            'coordinates' => [10, 20],
-                            'labels' => [
-                                [
-                                    'label' => 123321,
-                                    'annotator' => '123abc',
-                                    'confidence' => 1.0,
-                                    'created-at' => '2022-02-10 09:47:00',
-                                ],
-                            ],
-                        ],
-                    ],
-                ],
-            ],
-        ];
-
-        [$volume, $ifdo] = $this->setUpIfdo($merge);
-
-        $a = ImageAnnotationTest::create([
-            'image_id' => $image->id,
-            'points' => [150, 150],
-            'shape_id' => Shape::pointId(),
-        ]);
-        $al = ImageAnnotationLabelTest::create([
-            'label_id' => $label->id,
-            'user_id' => $user->id,
-            'annotation_id' => $a->id,
-        ]);
-
-        $image->volume_id = $volume->id;
-        $image->save();
-
-        $generator = new ImageIfdoReportGeneratorStub([
-            'stripIfdo' => true,
-        ]);
-        $generator->setSource($volume);
-        $generator->generateReport('my/path');
-
-        $expect = [
-            'image-set-header' => [
-                'image-set-handle' => '20.500.12085/test-example',
-                'image-set-name' => 'My Cool Volume',
-                'image-set-uuid' => 'd7546c4b-307f-4d42-8554-33236c577450',
-                'image-acquisition' => 'image',
-                'image-annotation-creators' => [
-                    [
-                        'id' => $user->uuid,
-                        'name' => "{$user->firstname} {$user->lastname}",
-                        'type' => 'expert',
-                    ],
-                ],
-                'image-annotation-labels' => [
-                    [
-                        'id' => $al->label_id,
-                        'name' => $al->label->name,
-                    ],
-                ],
-            ],
-            'image-set-items' => [
-                $al->annotation->image->filename => [
-                    'image-annotation-geometry-types' => ['single-pixel'],
-                    'image-annotations' => [
-                        [
-                            'coordinates' => $al->annotation->points,
-                            'labels' => [
-                                [
-                                    'label' => $al->label_id,
-                                    'annotator' => $user->uuid,
-                                    'confidence' => $al->confidence,
-                                    'created-at' => (string) $al->created_at,
-                                ],
-                            ],
-                        ],
-                    ],
-                ],
-            ],
-        ];
-
-        $this->assertEquals($expect, $generator->yaml);
-    }
-
-    public function testStripIfdoArray()
-    {
-        $label = LabelTest::create();
-        $user = UserTest::create();
-
-        $image = ImageTest::create();
-
-        $merge = [
-            'image-set-header' => [
-                'image-annotation-creators' => [
-                    [
-                        'id' => '123abc',
-                        'name' => "Test User",
-                        'type' => 'expert',
-                    ],
-                ],
-                'image-annotation-labels' => [
-                    [
-                        'id' => 123321,
-                        'name' => 'Test Label',
-                    ],
-                ],
-            ],
-            'image-set-items' => [
-                $image->filename => [[
+                $video->filename => [[
                     'image-annotation-geometry-types' => ['bounding-box'],
                     'image-annotations' => [
                         [
@@ -994,21 +825,22 @@ class ImageIfdoReportGeneratorTest extends TestCase
 
         [$volume, $ifdo] = $this->setUpIfdo($merge);
 
-        $a = ImageAnnotationTest::create([
-            'image_id' => $image->id,
-            'points' => [150, 150],
+        $a = VideoAnnotationTest::create([
+            'video_id' => $video->id,
+            'points' => [[150, 150]],
+            'frames' => [100.0],
             'shape_id' => Shape::pointId(),
         ]);
-        $al = ImageAnnotationLabelTest::create([
+        $al = VideoAnnotationLabelTest::create([
             'label_id' => $label->id,
             'user_id' => $user->id,
             'annotation_id' => $a->id,
         ]);
 
-        $image->volume_id = $volume->id;
-        $image->save();
+        $video->volume_id = $volume->id;
+        $video->save();
 
-        $generator = new ImageIfdoReportGeneratorStub([
+        $generator = new VideoIfdoReportGeneratorStub([
             'stripIfdo' => true,
         ]);
         $generator->setSource($volume);
@@ -1019,7 +851,7 @@ class ImageIfdoReportGeneratorTest extends TestCase
                 'image-set-handle' => '20.500.12085/test-example',
                 'image-set-name' => 'My Cool Volume',
                 'image-set-uuid' => 'd7546c4b-307f-4d42-8554-33236c577450',
-                'image-acquisition' => 'image',
+                'image-acquisition' => 'video',
                 'image-annotation-creators' => [
                     [
                         'id' => $user->uuid,
@@ -1035,11 +867,12 @@ class ImageIfdoReportGeneratorTest extends TestCase
                 ],
             ],
             'image-set-items' => [
-                $al->annotation->image->filename => [[
+                $al->annotation->video->filename => [[
                     'image-annotation-geometry-types' => ['single-pixel'],
                     'image-annotations' => [
                         [
-                            'coordinates' => $al->annotation->points,
+                            'coordinates' => [[150, 150]],
+                            'frames' => [100.0],
                             'labels' => [
                                 [
                                     'label' => $al->label_id,
@@ -1063,61 +896,71 @@ class ImageIfdoReportGeneratorTest extends TestCase
 
         $label = LabelTest::create();
         $user = UserTest::create();
-        $image = ImageTest::create([
+        $video = VideoTest::create([
             'volume_id' => $volume->id,
         ]);
 
-        $a1 = ImageAnnotationTest::create([
-            'image_id' => $image->id,
+        $a1 = VideoAnnotationTest::create([
+            'video_id' => $video->id,
             'shape_id' => Shape::pointId(),
         ]);
-        $al1 = ImageAnnotationLabelTest::create([
+        $al1 = VideoAnnotationLabelTest::create([
             'label_id' => $label->id,
             'user_id' => $user->id,
             'annotation_id' => $a1->id,
         ]);
 
-        $a2 = ImageAnnotationTest::create([
-            'image_id' => $image->id,
+        $a2 = VideoAnnotationTest::create([
+            'video_id' => $video->id,
             'shape_id' => Shape::rectangleId(),
         ]);
-        $al2 = ImageAnnotationLabelTest::create([
+        $al2 = VideoAnnotationLabelTest::create([
             'label_id' => $label->id,
             'user_id' => $user->id,
             'annotation_id' => $a2->id,
         ]);
 
-        $a3 = ImageAnnotationTest::create([
-            'image_id' => $image->id,
+        $a3 = VideoAnnotationTest::create([
+            'video_id' => $video->id,
             'shape_id' => Shape::circleId(),
         ]);
-        $al3 = ImageAnnotationLabelTest::create([
+        $al3 = VideoAnnotationLabelTest::create([
             'label_id' => $label->id,
             'user_id' => $user->id,
             'annotation_id' => $a3->id,
         ]);
 
-        $a4 = ImageAnnotationTest::create([
-            'image_id' => $image->id,
+        $a4 = VideoAnnotationTest::create([
+            'video_id' => $video->id,
             'shape_id' => Shape::ellipseId(),
         ]);
-        $al4 = ImageAnnotationLabelTest::create([
+        $al4 = VideoAnnotationLabelTest::create([
             'label_id' => $label->id,
             'user_id' => $user->id,
             'annotation_id' => $a4->id,
         ]);
 
-        $a5 = ImageAnnotationTest::create([
-            'image_id' => $image->id,
+        $a5 = VideoAnnotationTest::create([
+            'video_id' => $video->id,
             'shape_id' => Shape::polygonId(),
         ]);
-        $al5 = ImageAnnotationLabelTest::create([
+        $al5 = VideoAnnotationLabelTest::create([
             'label_id' => $label->id,
             'user_id' => $user->id,
             'annotation_id' => $a5->id,
         ]);
 
-        $generator = new ImageIfdoReportGeneratorStub;
+        $a6 = VideoAnnotationTest::create([
+            'video_id' => $video->id,
+            'shape_id' => Shape::wholeFrameId(),
+        ]);
+        $al6 = VideoAnnotationLabelTest::create([
+            'label_id' => $label->id,
+            'user_id' => $user->id,
+            'annotation_id' => $a6->id,
+        ]);
+
+        $generator = new VideoIfdoReportGeneratorStub;
         $generator->setSource($volume);
         $generator->generateReport('my/path');
 
@@ -1126,7 +969,7 @@ class ImageIfdoReportGeneratorTest extends TestCase
                 'image-set-handle' => '20.500.12085/test-example',
                 'image-set-name' => 'My Cool Volume',
                 'image-set-uuid' => 'd7546c4b-307f-4d42-8554-33236c577450',
-                'image-acquisition' => 'image',
+                'image-acquisition' => 'video',
                 'image-annotation-creators' => [
                     [
                         'id' => $user->uuid,
@@ -1142,11 +985,12 @@ class ImageIfdoReportGeneratorTest extends TestCase
                 ],
             ],
             'image-set-items' => [
-                $image->filename => [
-                    'image-annotation-geometry-types' => ['single-pixel', 'polygon', 'bounding-box'],
+                $video->filename => [[
+                    'image-annotation-geometry-types' => ['single-pixel', 'polygon', 'bounding-box', 'whole-image'],
                     'image-annotations' => [
                         [
                             'coordinates' => $a1->points,
+                            'frames' => $a1->frames,
                             'labels' => [
                                 [
                                     'label' => $label->id,
@@ -1158,6 +1002,7 @@ class ImageIfdoReportGeneratorTest extends TestCase
                         ],
                         [
                             'coordinates' => $a2->points,
+                            'frames' => $a2->frames,
                             'labels' => [
                                 [
                                     'label' => $label->id,
@@ -1169,6 +1014,7 @@ class ImageIfdoReportGeneratorTest extends TestCase
                         ],
                         [
                             'coordinates' => $a3->points,
+                            'frames' => $a3->frames,
                             'labels' => [
                                 [
                                     'label' => $label->id,
@@ -1180,6 +1026,7 @@ class ImageIfdoReportGeneratorTest extends TestCase
                         ],
                         [
                             'coordinates' => $a4->points,
+                            'frames' => $a4->frames,
                             'labels' => [
                                 [
                                     'label' => $label->id,
@@ -1191,6 +1038,7 @@ class ImageIfdoReportGeneratorTest extends TestCase
                         ],
                         [
                             'coordinates' => $a5->points,
+                            'frames' => $a5->frames,
                             'labels' => [
                                 [
                                     'label' => $label->id,
@@ -1200,8 +1048,20 @@ class ImageIfdoReportGeneratorTest extends TestCase
                                 ],
                             ],
                         ],
+                        [
+                            'coordinates' => $a6->points,
+                            'frames' => $a6->frames,
+                            'labels' => [
+                                [
+                                    'label' => $label->id,
+                                    'annotator' => $user->uuid,
+                                    'confidence' => $al6->confidence,
+                                    'created-at' => (string) $al6->created_at,
+                                ],
+                            ],
+                        ],
                     ],
-                ],
+                ]],
             ],
         ];
 
@@ -1209,7 +1069,7 @@ class ImageIfdoReportGeneratorTest extends TestCase
     }
 }
 
-class ImageIfdoReportGeneratorStub extends ImageIfdoReportGenerator
+class VideoIfdoReportGeneratorStub extends VideoIfdoReportGenerator
 {
     public $yaml;
